# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WikiLLM API
  description: |-
    HTTP API for [wiki-llm](https://gist.github.com/karpathy/442a6bf555914893e9891c11519de94f)
    folders. Designed to run alongside Obsidian, git, LiveSync, and other sync tools while
    serving multiple users and agents from the same markdown wiki.

    ## Project links

    - [GitHub repository](https://github.com/LukasParke/wikillm-api)
    - [Docker image](https://github.com/LukasParke/wikillm-api/pkgs/container/wikillm-api)
    - [Deployment guide](https://github.com/LukasParke/wikillm-api#deployment)
    - [Architecture docs](https://github.com/LukasParke/wikillm-api/blob/main/docs/ARCHITECTURE.md)

    ## Features

    - Self-hosted knowledge base service for humans and agents
    - Filesystem remains the source of truth; the index is a rebuildable cache
    - Atomic writes with optimistic concurrency control
    - Multi-source API key attribution with OKF actor convention
    - Hybrid retrieval: full-text + vector + recency decay, RRF fusion (K=60), optional rerank
    - LLM answering with citations (planner → executor → synthesis)
    - SSE and WebSocket live change feeds
    - Batch ingestion of sources and wiki pages
    - Connector framework (git, web, GitHub)
    - MCP server exposing LLM-free retrieval tools
    - Projects scoping with per-key RBAC (read/write/admin)
    - Google OKF v0.2 conformance: validation, trust tiers, provenance stamping
  version: 0.1.0
  contact:
    name: WikiLLM API
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key passed as a Bearer token. Keys are configured via the API_KEYS environment variable.
  schemas:
    Page:
      type: object
      required:
        - rel_path
        - abs_path
        - title
        - summary
        - frontmatter
        - body
        - word_count
        - outgoing_links
        - hash
        - mtime
        - updated_at
        - updated_by
      properties:
        rel_path:
          type: string
          example: wiki/entities/OpenAI.md
        abs_path:
          type: string
          example: /wiki/wiki/entities/OpenAI.md
        title:
          type: string
          nullable: true
          example: OpenAI
        summary:
          type: string
          nullable: true
        frontmatter:
          type: object
          example:
            tags:
              - ai
            updated_at: "2026-06-22T21:30:25.762Z"
            updated_by: agent-codex
        body:
          type: string
          example: "# OpenAI\n\nOpenAI is an AI research company."
        word_count:
          type: integer
          example: 7
        outgoing_links:
          type: array
          items:
            type: string
          example:
            - "[[Sam Altman]]"
        hash:
          type: string
          description: SHA-256 of the file content
          example: c7bcc0f1215169cee45e9eed89792a5f955aa63621405a56a0fa8ff2de424eea
        mtime:
          type: number
          description: File modification time in milliseconds
        updated_at:
          type: string
          format: date-time
          nullable: true
        updated_by:
          type: string
          nullable: true
          example: agent-codex
    PageList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Page"
        nextCursor:
          type: string
          nullable: true
    PageWriteInput:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Markdown body of the page
        frontmatter:
          type: object
          description: |
            YAML frontmatter fields. updated_at and updated_by are auto-injected if absent.
        ifMatch:
          type: string
          nullable: true
          description: |
            SHA-256 hash of the current file content. Required for updates; omit for creates.
    PageWriteResult:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        page:
          $ref: "#/components/schemas/Page"
        operationId:
          type: string
    SourceFile:
      type: object
      required:
        - rel_path
        - abs_path
        - content_type
        - size
        - hash
        - mtime
      properties:
        rel_path:
          type: string
          example: raw/articles/example.md
        abs_path:
          type: string
        content_type:
          type: string
          nullable: true
          example: text/markdown
        size:
          type: integer
          example: 1234
        hash:
          type: string
        mtime:
          type: number
    SourceList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SourceFile"
        nextCursor:
          type: string
          nullable: true
    Conflict:
      type: object
      required:
        - hash
        - content
      properties:
        hash:
          type: string
        content:
          type: string
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: conflict
        message:
          type: string
        current:
          $ref: "#/components/schemas/Conflict"
        existingHash:
          type: string
...
    GraphNode:
      type: object
      required:
        - rel_path
      properties:
        rel_path:
          type: string
        title:
          type: string
          nullable: true
        exists:
          type: boolean
          description: False for broken links tolerated per the OKF spec
    GraphEdge:
      type: object
      required:
        - src
        - dst
      properties:
        src:
          type: string
        dst:
          type: string
    GraphResponse:
      type: object
      required:
        - nodes
        - edges
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
...
    BatchPageOperation:
      type: object
      required:
        - rel_path
      properties:
        rel_path:
          type: string
        content:
          type: string
          description: New markdown body; omit for delete operations
        frontmatter:
          type: object
          description: YAML frontmatter fields merged onto the write
        ifMatch:
          type: string
          nullable: true
          description: Expected SHA-256 hash; any mismatch aborts the whole batch
        delete:
          type: boolean
          description: Delete the page instead of writing it
    BatchPageInput:
      type: object
      required:
        - operations
      properties:
        operations:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/BatchPageOperation"
    BatchOperationResult:
      type: object
      required:
        - rel_path
        - success
      properties:
        rel_path:
          type: string
        success:
          type: boolean
        error:
          type: string
          nullable: true
        conflict:
          $ref: "#/components/schemas/Conflict"
    BatchPageResult:
      type: object
      required:
        - success
        - results
      properties:
        success:
          type: boolean
        results:
          type: array
          items:
            $ref: "#/components/schemas/BatchOperationResult"
...
paths:
...
  /v1/pages:
    get:
      tags:
        - Pages
      summary: List wiki pages
      description: List pages from the SQLite cache with optional pagination.
      parameters:
        - name: folder
          in: query
          schema:
            type: string
            default: wiki
          description: Relative folder to list from
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
          description: Path to start after for pagination
      responses:
        "200":
          description: List of pages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageList"
  /v1/pages/{rel_path}:
    parameters:
      - name: rel_path
        in: path
        required: true
        schema:
          type: string
        description: Relative path of the page, e.g. wiki/entities/OpenAI.md
        example: wiki/entities/OpenAI.md

    get:
      tags:
        - Pages
      summary: Read a wiki page
      responses:
        "200":
          description: Page found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          description: Page not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags:
        - Pages
      summary: Create or update a wiki page
      description: |
        Writes a page atomically. If the page exists, ifMatch must match the current
        SHA-256 hash or a 409 conflict is returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageWriteInput"
      responses:
        "200":
          description: Page written successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageWriteResult"
        "409":
          description: Optimistic concurrency conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - Pages
      summary: Delete a wiki page
      description: |
        Deletes a page. If an If-Match header is provided and does not match the current
        hash, a 409 conflict is returned.
      parameters:
        - name: If-Match
          in: header
          schema:
            type: string
          description: SHA-256 hash of the current content
      responses:
        "200":
          description: Page deleted
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - operationId
                properties:
                  success:
                    type: boolean
                  operationId:
                    type: string
        "404":
          description: Page not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Optimistic concurrency conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/pages/batch:
    post:
      tags:
        - Pages
      summary: Apply many page writes atomically
      description: |
        All-or-nothing batch write (max 1000 operations). Every operation is
        preflight-checked against its ifMatch hash before anything is written;
        a single conflict aborts the whole batch and nothing is mutated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchPageInput"
      responses:
        "200":
          description: All operations applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchPageResult"
        "409":
          description: Preflight conflict; nothing was written
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                  - results
                properties:
                  error:
                    type: string
                    example: conflict
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/BatchOperationResult"
  /v1/sources:
    get:
      tags:
        - Sources
      summary: List raw sources
      parameters:
        - name: folder
          in: query
          schema:
            type: string
            default: raw
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceList"
  /v1/sources/{rel_path}:
    parameters:
      - name: rel_path
        in: path
        required: true
        schema:
          type: string
        description: Relative path of the source, e.g. raw/articles/example.md
        example: raw/articles/example.md

    get:
      tags:
        - Sources
      summary: Read source metadata
      responses:
        "200":
          description: Source metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceFile"
        "404":
          description: Source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Sources
      summary: Upload a raw source
      description: |
        Sources are write-once by default. Use ?force=true to overwrite an existing source.
      parameters:
        - name: force
          in: query
          schema:
            type: boolean
            default: false
          description: Overwrite if source already exists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
          application/octet-stream:
            schema:
              type: string
              format: binary
          text/plain:
            schema:
              type: string
      responses:
        "201":
          description: Source created or overwritten
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceFile"
        "409":
          description: Source already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - Sources
      summary: Delete a raw source
      responses:
        "200":
          description: Source deleted
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
        "404":
          description: Source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/pages/{rel_path}/raw:
    parameters:
      - name: rel_path
        in: path
        required: true
        schema:
          type: string
        description: Relative path of the page, e.g. wiki/entities/OpenAI.md
        example: wiki/entities/OpenAI.md

    get:
      tags:
        - Pages
      summary: Read a page's raw markdown
      description: Returns the page body as text/markdown with an ETag header.
      responses:
        "200":
          description: Raw markdown body
          headers:
            ETag:
              schema:
                type: string
          content:
            text/markdown:
              schema:
                type: string
        "404":
          description: Page not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/sources/{rel_path}/content:
    parameters:
      - name: rel_path
        in: path
        required: true
        schema:
          type: string
        description: Relative path of the source, e.g. raw/articles/example.md
        example: raw/articles/example.md

    get:
      tags:
        - Sources
      summary: Download a source's original bytes
      description: Streams the stored file with its detected Content-Type and an ETag header.
      responses:
        "200":
          description: Original source bytes
          headers:
            Content-Type:
              schema:
                type: string
              description: Detected media type of the stored file
            ETag:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/graph:
    get:
      tags:
        - Graph
      summary: Link graph neighbors (query-param form)
      description: Equivalent to /v1/graph/{rel_path}; accepts the same parameters.
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Relative path of the page
        - name: depth
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 3
            default: 1
        - name: format
          in: query
          schema:
            type: string
            enum:
              - json
              - dot
            default: json
          description: json (default) or dot (text/vnd.graphviz)
      responses:
        "200":
          description: Graph neighborhood
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphResponse"
            text/vnd.graphviz:
              schema:
                type: string
        "404":
          description: Page not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/graph/{rel_path}:
    parameters:
      - name: rel_path
        in: path
        required: true
        schema:
          type: string
        description: Relative path of the page, e.g. wiki/entities/OpenAI.md
        example: wiki/entities/OpenAI.md

    get:
      tags:
        - Graph
      summary: Link graph neighbors
      description: Returns nodes and edges around a page, following wiki links. Accepts format=json (default) or format=dot (text/vnd.graphviz).
      parameters:
        - name: depth
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 3
            default: 1
        - name: format
          in: query
          schema:
            type: string
            enum:
              - json
              - dot
            default: json
          description: json (default) or dot (text/vnd.graphviz)
      responses:
        "200":
          description: Graph neighborhood
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphResponse"
            text/vnd.graphviz:
              schema:
                type: string
              description: DOT output when format=dot
        "404":
          description: Page not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...